add screenshot and several locations to ui tests. make text not editable and show user location from the start.
XCUITests should now pass locally on Simulators and Real Device as well as on Sauce EU and US RDC.
Had to bake default Sauce locations since it seems that the default location on XCUITest is not used.

Made location text not editable for easy copy and pasting. Current location dot now shows up on first launch.